% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{covDataAccess}
\alias{covDataAccess}
\title{Differentially Private Covariance Data Access Function}
\usage{
covDataAccess(x1, x2, lower.bound1, upper.bound1, lower.bound2, upper.bound2)
}
\arguments{
\item{x1, x2}{Numeric vectors.}

\item{lower.bound1, lower.bound2}{Real numbers giving the lower bounds of x1
and x2, respectively.}

\item{upper.bound1, upper.bound2}{Real numbers giving the upper bounds of x1
and x2, respectively.}
}
\value{
List of the true covariance, and the bounded and unbounded
sensitivities.
}
\description{
This function performs the data access step in the computation of a
differentially private covariance. The true values are computed using
\code{\link[stats]{cov}}, while the bounded and unbounded sensitivities are
computed according to the theoretical values from \link{TODO: CITE PAPER HERE}.
}
\examples{
covDataAccess(c(1,4,3,2), c(-2,-3,-4,-1), 0, 5, -5, 0)

}
\keyword{internal}
