% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{covDP}
\alias{covDP}
\title{Differentially Private Covariance}
\usage{
covDP(
  x1,
  x2,
  eps,
  which.sensitivity = "bounded",
  lower.bound1 = NULL,
  upper.bound1 = NULL,
  lower.bound2 = NULL,
  upper.bound2 = NULL,
  mechanism = "laplace",
  delt = NULL,
  type.DP = "pDP"
)
}
\arguments{
\item{x1, x2}{Numeric vectors.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded sensitivities. If 'unbounded',
returns result plus noise based on unbounded sensitivities. If 'both',
returns result based on both methods. Note that if 'both' is chosen, each
result individually satisfies differential privacy at level eps, but may
not do so collectively. Care must be taken not to violate differential
privacy in this case.}

\item{lower.bound1, lower.bound2}{Real numbers giving the lower bounds of x1
and x2, respectively. If not given, they are computed from the data to be
the min value of x1 and x2, respectively. Note this computation may result
in additional privacy loss.}

\item{upper.bound1, upper.bound2}{Real numbers giving the upper bounds of x1
and x2, respectively. If not given, they are computed from the data to be
the max value of x1 and x2, respectively. Note this computation may result
in additional privacy loss.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'laplace',
'gaussian'}. See \code{\link{laplaceMechanism}} and
\code{\link{gaussianMechanism}} for a description of the supported
mechanisms.}

\item{delt}{Positive real number defining the delta privacy parameter
(necessary if using Gaussian mechanism).}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism. Can be either probabilistic DP ('pDP') or
approximate DP ('aDP'). See \code{\link{gaussianMechanism}} for a more
detailed description of this parameter. Only used if mechanism is
'gaussian'.}
}
\value{
List of bounded and/or unbounded sanitized covariances.
}
\description{
This function computes the differentially private covariance of a pair of
vectors at user-specified levels of epsilon and delta.
}
\examples{
covDP(c(1,4,-2,8,-6),c(1,3,2,2,4),1,which.sensitivity='bounded',
  lower.bound1=-10,upper.bound1=10,lower.bound2=0,upper.bound2=5,
  mechanism='laplace')
covDP(c(1,4,-2,8,-6),c(1,3,2,2,4),1,which.sensitivity='unbounded',
  lower.bound1=-10,upper.bound110,lower.bound2=0,upper.bound2=5,
  mechanism='gaussian',delt=0.5,type.DP='aDP')

}
