% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mechanisms.R
\name{exponentialMechanism}
\alias{exponentialMechanism}
\title{Exponential Mechanism}
\usage{
exponentialMechanism(
  utility,
  eps,
  bounded.sensitivities = NULL,
  unbounded.sensitivities = NULL,
  which.sensitivity = "bounded",
  measure = NULL,
  candidates = NULL
)
}
\arguments{
\item{utility}{Numeric vector giving the utilities of the possible values.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{bounded.sensitivities}{Real number corresponding to the bounded global
sensitivity of the statistic. This is defined to be the greatest amount by
which the statistic could change in value by changing one entry in the
dataset (i.e. the total number of elements in the dataset remain the same).
This value can only be NULL if which.sensitivity is 'unbounded'.}

\item{unbounded.sensitivities}{Real number corresponding to the unbounded
global sensitivity of the statistic. This is defined to be the greatest
amount by which the statistic could change in value by adding/removing one
entry of the dataset (i.e. the total number of elements in the dataset
increases/decreases by one). This value can only be NULL if
which.sensitivity is 'bounded'.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded sensitivities. If 'unbounded',
returns result plus noise based on unbounded sensitivities. If 'both',
returns result based on both methods. Note that if 'both' is chosen, each
result individually satisfies differential privacy at level eps, but may
not do so collectively. Care must be taken not to violate differential
privacy in this case.}

\item{measure}{Optional numeric vector of scaling measures for each element
of utility. Should be same size as utility. Defaults to uniform scaling.}

\item{candidates}{Optional vector of candidates of same size as utility. If
given, the function returns the candidate at the selected index rather than
the index itself.}
}
\value{
List of indices (or values if candidates given) selected by the
mechanism based on bounded and/or unbounded sensitivities.
}
\description{
TODO: Add description.
}
\examples{
exponentialMechanism(c(0,1,2,3,2,1,0), 1, bounded.sensitivities=1,
  which.sensitivity='bounded')
exponentialMechanism(c(0,1,2,3,2,1,0), 1, unbounded.sensitivities=.5,
  which.sensitivity='unbounded', measure=c(1,1,2,1,2,1,1),
  candidates=c('a','b','c','d','e','f','g'))

}
