% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mechanisms.R
\name{GaussianMechanism}
\alias{GaussianMechanism}
\title{Gaussian Mechanism}
\usage{
GaussianMechanism(
  true.values,
  eps,
  delta,
  bounded.sensitivities = NULL,
  unbounded.sensitivities = NULL,
  which.sensitivity = "bounded",
  type.DP = "aDP",
  alloc.proportions = NULL
)
}
\arguments{
\item{true.values}{Real number or numeric vector corresponding to the true
value(s) of the desired function(s).}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{delta}{Positive real number defining the delta privacy parameter.}

\item{bounded.sensitivities}{Real number or numeric vector corresponding to
the l2-global sensitivity(-ies) of the function(s) based on bounded
differential privacy \insertCite{Kifer2011}{DPpack}. This is defined to be
the greatest amount by which the function could change in l2-norm by
changing one entry in the dataset (i.e. the total number of elements in the
dataset remain the same). This value can only be NULL if which.sensitivity
is 'unbounded'. This value must be of length 1 or of the same length as
true.values. If it is of length 1 and true.values is a vector, this
indicates that the given sensitivity applies simultaneously to all elements
of true.values and that the privacy budget need not be allocated
(alloc.proportions is unused in this case). If it is of length larger than
1, this indicates that the Gaussian mechanism is being applied to each
element of true.values individually. In this case the privacy budget is
divided using alloc.proportions.}

\item{unbounded.sensitivities}{Real number or numeric vector corresponding to
the l2-global sensitivity(-ies) of the function(s) based on unbounded
differential privacy \insertCite{Kifer2011}{DPpack}. This is defined to be
the greatest amount by which the function could change in l2-norm by
adding/removing one entry of the dataset (i.e. the total number of elements
in the dataset increases/decreases by one). This value can only be NULL if
which.sensitivity is 'bounded'. This value must be of length 1 or of the
same length as true.values. If it is of length 1 and true.values is a
vector, this indicates that the given sensitivity applies simultaneously to
all elements of true.values and that the privacy budget need not be
allocated (alloc.proportions is unused in this case). If it is of length
larger than 1, this indicates that the Gaussian mechanism is being applied
to each element of true.values individually. In this case the privacy
budget is divided using alloc.proportions.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively. Care must be taken not to violate differential privacy in
this case.}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism. Can be either 'pDP' for probabilistic DP
\insertCite{Liu2019a}{DPpack} or 'aDP' for approximate DP
\insertCite{DPtextbook}{DPpack}. Note that if 'aDP' is chosen, epsilon must
be strictly less than 1.}

\item{alloc.proportions}{Numeric vector giving the allocation proportions of
epsilon (and delta if relevant) to the function values. For example, if
true.values is of length two and alloc.proportions = c(.75, .25), then 75\%
of the privacy budget eps (and delta) is allocated to the noise computation
for the first element of true.values, and the remaining 25\% is allocated to
the noise computation for the second element of true.values. This ensures
(eps, delta)-level privacy across all computations. By default, it
distributes eps and delta evenly among the calculations. Input does not
need to be normalized, meaning alloc.proportions = c(3,1) produces the same
result as the example above.}
}
\value{
Sanitized function values based on the bounded and/or unbounded
definitions of differential privacy, sanitized via the Gaussian mechanism.
}
\description{
This function implements the Gaussian mechanism for differential privacy by
adding noise to the true value of a function according to specified values of
epsilon, delta, and l2-global sensitivity. Sensitivity calculated based
either on bounded or unbounded differential privacy can be used
\insertCite{Kifer2011}{DPpack}. If true.values is a vector, the provided
epsilon and delta are divided such that (epsilon, delta)-level differential
privacy is satisfied across all function values. If desired, the user can
specify how to divide epsilon and delta among the function values using
alloc.proportions.
}
\examples{
GaussianMechanism(5, 1, .01, bounded.sensitivities=0.5,
  which.sensitivity='bounded',type.DP='pDP')
GaussianMechanism(c(5,3), .5, .01, unbounded.sensitivities=c(1,1),
  which.sensitivity='unbounded', alloc.proportions=c(.4,.6))

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}

\insertRef{Liu2019a}{DPpack}

\insertRef{DPtextbook}{DPpack}
}
