% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{generate.sampling}
\alias{generate.sampling}
\title{Generator for Sampling Distribution Function for Gaussian Kernel}
\usage{
generate.sampling(gamma)
}
\arguments{
\item{gamma}{Positive real number for the Gaussian (radial) kernel parameter.}
}
\value{
Sampling function for the Gaussian kernel of form required by
\code{\link{svmDP}}.
}
\description{
This function generates and returns a sampling function corresponding to the
Fourier transform of a Gaussian kernel with parameter gamma
\insertCite{chaudhuri2011}{DPpack} of form needed for \code{\link{svmDP}}.
}
\examples{
  gamma <- 1
  sample <- generate.sampling(gamma)
  d <- 5
  sample(d)

}
\references{
\insertRef{chaudhuri2011}{DPpack}
}
\keyword{internal}
