% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{h.linear}
\alias{h.linear}
\title{Linear Hypothesis Function}
\usage{
h.linear(X, coeff)
}
\arguments{
\item{X}{Matrix of data.}

\item{coeff}{Vector or matrix of coefficients or weights.}
}
\value{
Matrix of values of the linear hypothesis function corresponding to
each row of X.
}
\description{
This function implements the linear hypothesis function used for linear SVM
and linear regression in the form required by
\code{\link{EmpiricalRiskMinimizationDP.CMS}} and
\code{\link{EmpiricalRiskMinimizationDP.KST}}.
}
\examples{
  X <- matrix(c(1,2,3,4,5,6),nrow=2)
  coeff <- c(0.5,-1,2)
  h.linear(X,coeff)

}
\keyword{internal}
