% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{histogramDataAccess}
\alias{histogramDataAccess}
\title{Differentially Private Histogram Data Access Function}
\usage{
histogramDataAccess(x, breaks)
}
\arguments{
\item{x}{Numeric vector.}

\item{breaks}{Identical to the argument with the same name from
\code{\link[graphics]{hist}}.}
}
\value{
List of the true histogram, and the bounded and unbounded
sensitivities.
}
\description{
This function performs the data access step in the computation of a
differentially private histogram. The true values are computed using
\code{\link[graphics]{hist}}, while the bounded and unbounded sensitivities are
computed according to the theoretical values from \link{TODO: CITE PAPER HERE}.
}
\examples{
histogramDataAccess(c(1,4,3,2,3), 'Sturges')

}
\keyword{internal}
