% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{histogramDP}
\alias{histogramDP}
\title{Differentially Private Histogram}
\usage{
histogramDP(
  x,
  eps,
  breaks = "Sturges",
  normalize = FALSE,
  which.sensitivity = "bounded",
  lower.bound = NULL,
  upper.bound = NULL,
  mechanism = "laplace",
  delt = NULL,
  type.DP = "pDP",
  allow.negative = FALSE
)
}
\arguments{
\item{x}{Numeric vector from which the histogram will be formed.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{breaks}{Identical to the argument with the same name from
\code{\link[graphics]{hist}}.}

\item{normalize}{Logical value. If FALSE (default), returned histogram counts
correspond to frequencies. If TRUE, returned histogram counts correspond to
densities (i.e. area of histogram is one).}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively. Care must be taken not to violate differential privacy in
this case.}

\item{lower.bound}{Real number giving the lower bound of x. If not given, it
is computed from the data to be the min value of x. Note this computation
may result in additional privacy loss.}

\item{upper.bound}{Real number giving the upper bound of x. If not given, it
is computed from the data to be the max value of x. Note this computation
may result in additional privacy loss.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'laplace',
'gaussian'}. See \code{\link{laplaceMechanism}} and
\code{\link{gaussianMechanism}} for a description of the supported
mechanisms.}

\item{delt}{Positive real number defining the delta privacy parameter
(necessary if using Gaussian mechanism).}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism. Can be either probabilistic DP ('pDP') or
approximate DP ('aDP'). See \code{\link{gaussianMechanism}} for a more
detailed description of this parameter. Only used if mechanism is
'gaussian'.}

\item{allow.negative}{Logical value. If FALSE (default), any negative values
in the sanitized histogram due to the added noise will be set to 0. If
TRUE, the negative values (if any) will be returned.}
}
\value{
A list of the sanitized histograms based on the bounded and/or
unbounded definitions of differential privacy.
}
\description{
This function computes a differentially private histogram from a vector at
user-specified levels of epsilon and delta. A histogram object is returned
with sanitized values for the counts for easy plotting (see examples).
}
\examples{
result <- histogramDP(c(1,1,-2,8,-6),1,which.sensitivity='bounded',
  lower.bound=-10,upper.bound=10,mechanism='laplace')
plot(result$Bounded)
result <- histogramDP(c(1,1,-2,8,-6),1,normalize=TRUE,
  which.sensitivity='unbounded',lower.bound=-10,upper.bound=10,
  mechanism='gaussian',delt=0.5,type.DP='aDP',allow.negative=FALSE)
plot(result$Unbounded)

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}
}
