% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{loss.cross.entropy}
\alias{loss.cross.entropy}
\title{Cross Entropy Loss Function}
\usage{
loss.cross.entropy(y.hat, y)
}
\arguments{
\item{y.hat}{Vector or matrix of estimated labels.}

\item{y}{Vector or matrix of true labels.}
}
\value{
Vector or matrix of the cross entropy loss for each element of y.hat
and y.
}
\description{
This function implements cross entropy loss used for logistic regression in
the form required by \code{\link{EmpiricalRiskMinimizationDP.CMS}}.
}
\examples{
  y.hat <- c(0.1, 0.88, 0.02)
  y <- c(0, 1, 0)
  loss.cross.entropy(y.hat,y)

}
\keyword{internal}
