% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{loss.gr.squared.error}
\alias{loss.gr.squared.error}
\title{Squared error Loss Function Gradient}
\usage{
loss.gr.squared.error(y.hat, y)
}
\arguments{
\item{y.hat}{Vector or matrix of estimated values.}

\item{y}{Vector or matrix of true values.}
}
\value{
Vector or matrix of the squared error loss gradient for each element
of y.hat and y.
}
\description{
This function implements the squared error loss gradient with respect to
y.hat used for linear regression in the form required by
\code{\link{EmpiricalRiskMinimizationDP.KST}}.
}
\examples{
  y.hat <- c(0.1, 0.88, 0.02)
  y <- c(-0.1, 1, .2)
  loss.gr.squared.error(y.hat,y)

}
\keyword{internal}
