% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{loss.squared.error}
\alias{loss.squared.error}
\title{Squared Error Loss Function}
\usage{
loss.squared.error(y.hat, y)
}
\arguments{
\item{y.hat}{Vector or matrix of estimated labels.}

\item{y}{Vector or matrix of true labels.}
}
\value{
Vector or matrix of the squared error loss for each element of y.hat
and y.
}
\description{
This function implements squared error loss used for linear regression in
the form required by \code{\link{EmpiricalRiskMinimizationDP.KST}}.
}
\examples{
  y.hat <- c(0.1, 0.88, 0.02)
  y <- c(0, 1, 0)
  loss.squared.error(y.hat,y)

}
\keyword{internal}
