% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{mapXy.sigmoid}
\alias{mapXy.sigmoid}
\title{Sigmoid Map Function}
\usage{
mapXy.sigmoid(X, coeff)
}
\arguments{
\item{X}{Matrix of data.}

\item{coeff}{Vector or matrix of coefficients or weights.}
}
\value{
Matrix of values of the sigmoid function corresponding to each row of
X.
}
\description{
This function implements the sigmoid map function from data X to labels y
used for logistic regression in the form required by
\code{\link{EmpiricalRiskMinimizationDP.CMS}}.
}
\examples{
  X <- matrix(c(1,2,3,4,5,6),nrow=2)
  coeff <- c(0.5,-1,2)
  mapXy.sigmoid(X,coeff)

}
\keyword{internal}
