% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{meanDataAccess}
\alias{meanDataAccess}
\title{Differentially Private Mean Data Access Function}
\usage{
meanDataAccess(x, lower.bounds, upper.bounds)
}
\arguments{
\item{x}{Numeric vector, matrix, or data frame. Means taken over columns
(when applicable).}

\item{lower.bounds}{Numeric vector of lower bounds on each column of x. The
length of lower.bounds must match the number of columns of x (length 1 if x
is a vector).}

\item{upper.bounds}{Numeric vector of upper bounds on each column of x. The
length of upper.bounds must match the number of columns of x (length 1 if x
is a vector).}
}
\value{
List of the true mean and the sensitivities calculated based on
bounded and unbounded differential privacy.
}
\description{
This function performs the data access step in the computation of a
differentially private mean. The true values are computed using
\code{\link[base]{mean}}, while the sensitivities are calculated based on
bounded and unbounded differential privacy \insertCite{Kifer2011}{DPpack}
according to the theoretical values \insertCite{Liu2019b}{DPpack}.
}
\examples{
meanDataAccess(c(1,4,3,2), 0, 5)

}
\references{
\insertRef{Liu2019b}{DPpack}

\insertRef{Kifer2011}{DPpack}
}
\keyword{internal}
