% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{meanDP}
\alias{meanDP}
\title{Compute differentially private mean}
\usage{
meanDP(
  x,
  eps,
  which.sensitivity = "bounded",
  lower.bounds = NULL,
  upper.bounds = NULL,
  mechanism = "laplace",
  delt = NULL,
  type.DP = "pDP",
  alloc.proportions = NULL
)
}
\arguments{
\item{x}{The dataset over which the mean will be taken. Means taken over columns.}

\item{eps}{The epsilon privacy budget.}

\item{which.sensitivity}{Can be one of {'bounded', 'unbounded', 'both'}. If
'bounded' (default), returns result plus noise based on bounded
sensitivities. If 'unbounded', returns result plus noise based on
unbounded sensitivities.}

\item{lower.bounds}{The list of lower bounds (minimums) of the full dataset
from which x was taken. If NULL, it is computed to be the min value of
each column of x.}

\item{upper.bounds}{The list of upper bound (maximum) of the full dataset
from which x was taken. If NULL, it is computed to be the min value of
each column of x.}

\item{mechanism}{A string indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported
{'laplace', 'gaussian'}.}

\item{delt}{The delta privacy parameter (necessary if using Gaussian mechanism).}

\item{type.DP}{The type of differential privacy desired for Gaussian mechanism.
Can be either probabilistic DP ('pDP') or approximate DP ('aDP').}

\item{alloc.proportions}{This gives the proportional allocation of epsilon (and
delta) to the statistics. By default, it distributes eps and delt
evenly among the calculations. Input does not need to be normalized.}
}
\value{
List of bounded and/or unbounded sanitized means.
}
\description{
TODO: Add description.
}
\examples{
meanDP(c(1,4,-2,8,-6),1,'bounded',-10,10,'laplace')
meanDP(c(1,4,-2,8,-6),1,'unbounded',-10,10,'gaussian',0.5,'aDP')

}
