% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{medianDP}
\alias{medianDP}
\title{Differentially Private Median}
\usage{
medianDP(
  x,
  eps,
  which.sensitivity = "bounded",
  lower.bound = NULL,
  upper.bound = NULL,
  mechanism = "exponential",
  delta = NULL
)
}
\arguments{
\item{x}{Numeric vector of which the median will be taken.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively and in composition. Care must be taken not to violate
differential privacy in this case.}

\item{lower.bound}{Real number giving the global or public lower bound of x.}

\item{upper.bound}{Real number giving the global or public upper bound of x.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'exponential'}.
See \code{\link{ExponentialMechanism}} for a description of the supported
mechanisms.}

\item{delta}{Nonnegative real number defining the delta privacy parameter. If
0, reduces to eps-DP and the Laplace mechanism is used.}
}
\value{
A list of the sanitized medians based on the bounded and/or unbounded
definitions of differential privacy.
}
\description{
This function computes the differentially private median of an input vector
at user-specified privacy levels of epsilon and delta.
}
\examples{
medianDP(c(1,1,-2,8,-6),1,which.sensitivity='bounded',
  lower.bound=-10,upper.bound=10,mechanism='exponential')
medianDP(c(1,1,-2,8,-6),1,which.sensitivity='unbounded',
  lower.bound=-10,upper.bound=10,mechanism='exponential')

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}

\insertRef{Smith2011a}{DPpack}
}
