% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{pooledCovDP}
\alias{pooledCovDP}
\title{Differentially Private Pooled Covariance}
\usage{
pooledCovDP(
  ...,
  eps = 1,
  which.sensitivity = "bounded",
  lower.bound1 = NULL,
  upper.bound1 = NULL,
  lower.bound2 = NULL,
  upper.bound2 = NULL,
  mechanism = "laplace",
  delt = NULL,
  type.DP = "pDP",
  approx.n.max = FALSE
)
}
\arguments{
\item{...}{Two or more matrices, each with two columns from which to compute
the pooled covariance.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded sensitivities. If 'unbounded',
returns result plus noise based on unbounded sensitivities. If 'both',
returns result based on both methods. Note that if 'both' is chosen, each
result individually satisfies differential privacy at level eps, but may
not do so collectively. Care must be taken not to violate differential
privacy in this case.}

\item{lower.bound1, lower.bound2}{Real numbers giving the lower bounds over
the first and second columns of all input data, respectively. If not given,
they are computed from the data to be the min value over all first and
second columns of the given data, respectively. Note this computation may
result in additional privacy loss.}

\item{upper.bound1, upper.bound2}{Real numbers giving the upper bounds over
the first and second columns of all input data, respectively. If not given,
they are computed from the data to be the max value over all first and
second columns of the given data, respectively. Note this computation may
result in additional privacy loss.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'laplace',
'gaussian'}. See \code{\link{laplaceMechanism}} and
\code{\link{gaussianMechanism}} for a description of the supported
mechanisms.}

\item{delt}{Positive real number defining the delta privacy parameter
(necessary if using Gaussian mechanism).}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism. Can be either probabilistic DP ('pDP') or
approximate DP ('aDP'). See \code{\link{gaussianMechanism}} for a more
detailed description of this parameter. Only used if mechanism is
'gaussian'.}

\item{approx.n.max}{Logical indicating whether to approximate n.max, which is
defined to be the length of the largest input vector. Approximation is best
if n.max is very large.}
}
\value{
List of bounded and/or unbounded sanitized pooled covariances.
}
\description{
This function computes the differentially private pooled covariance from two
or more two-column matrices of data at user-specified levels of epsilon and
delta.
}
\examples{
x1 <- matrix(c(1,4,-2,8,-6,-3),ncol=2)
x2 <- matrix(c(1,2,-5,7),ncol=2)
pooledVarDP(x1,x2,eps=1,which.sensitivity='bounded',
  lower.bound1=-10,upper.bound1=10,lower.bound2=-10,upper.bound2=10,
  mechanism='laplace')
pooledVarDP(x1,x2,eps=1,which.sensitivity='unbounded',
  lower.bound1=-10,upper.bound1=10,lower.bound2=-10,upper.bound2=10,
  mechanism='gaussian',delt=0.5,type.DP='aDP',approx.n.max=TRUE)

}
