% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{pooledVarDP}
\alias{pooledVarDP}
\title{Differentially Private Pooled Variance}
\usage{
pooledVarDP(
  ...,
  eps = 1,
  which.sensitivity = "bounded",
  lower.bound = NULL,
  upper.bound = NULL,
  mechanism = "laplace",
  delt = NULL,
  type.DP = "pDP",
  approx.n.max = FALSE
)
}
\arguments{
\item{...}{Two or more vectors from which to compute the pooled variance.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded sensitivities. If 'unbounded',
returns result plus noise based on unbounded sensitivities. If 'both',
returns result based on both methods. Note that if 'both' is chosen, each
result individually satisfies differential privacy at level eps, but may
not do so collectively. Care must be taken not to violate differential
privacy in this case.}

\item{lower.bound}{Real number giving the lower bound of the input data. If
not given, it is computed from the data to be the min value over all given
data. Note this computation may result in additional privacy loss.}

\item{upper.bound}{Real number giving the upper bound of the input data. If
not given, it is computed from the data to be the max value over all given
data. Note this computation may result in additional privacy loss.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'laplace',
'gaussian'}. See \code{\link{laplaceMechanism}} and
\code{\link{gaussianMechanism}} for a description of the supported
mechanisms.}

\item{delt}{Positive real number defining the delta privacy parameter
(necessary if using Gaussian mechanism).}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism. Can be either probabilistic DP ('pDP') or
approximate DP ('aDP'). See \code{\link{gaussianMechanism}} for a more
detailed description of this parameter. Only used if mechanism is
'gaussian'.}

\item{approx.n.max}{Logical indicating whether to approximate n.max, which is
defined to be the length of the largest input vector. Approximation is best
if n.max is very large.}
}
\value{
List of bounded and/or unbounded sanitized pooled variances.
}
\description{
This function computes the differentially private pooled variance from two or
more vectors of data at user-specified levels of epsilon and delta.
}
\examples{
pooledVarDP(c(1,4,-2,8,-6),c(1,2),c(-5,-7),eps=1,which.sensitivity='bounded',
  lower.bound=-10,upper.bound=10,mechanism='laplace')
pooledVarDP(c(1,4,-2,8,-6),c(1,2),c(-5,-7),eps=1,
  which.sensitivity='unbounded',lower.bound=-10,upper.bound=10,
  mechanism='gaussian',delt=0.5,type.DP='aDP',approx.n.max=TRUE)

}
