% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{quantileDataAccess}
\alias{quantileDataAccess}
\title{Differentially Private Quantile Data Access Function}
\usage{
quantileDataAccess(x, quant, lower.bound, upper.bound)
}
\arguments{
\item{x}{Numeric vector.}

\item{quant}{Real number between 0 and 1 indicating which quantile to return.}

\item{lower.bound}{Real number giving the lower bound of the input data.}

\item{upper.bound}{Real number giving the upper bound of the input data.}
}
\value{
List of a vector corresponding to the utility function, the sorted
and clipped vector of inputs, and the bounded and unbounded sensitivities.
}
\description{
This function performs the data access step in the computation of a
differentially private quantile. The utility vector is computed as in
\link{TODO:CITE THIS}, while the bounded and unbounded sensitivities are computed
according to the theoretical values from \link{TODO:PROVE THESE}.
}
\examples{
quantileDataAccess(c(1,1,-2,8,-6),.25,-10,10)

}
\keyword{internal}
