% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatModels.R
\name{regularizer.gr.l2}
\alias{regularizer.gr.l2}
\title{l2 Regularizer Gradient}
\usage{
regularizer.gr.l2(coeff)
}
\arguments{
\item{coeff}{Vector or matrix of coefficients or weights.}
}
\value{
Regularizer gradient value at the given coefficient vector.
}
\description{
This function implements the l2 regularizer gradient in the form required by
\code{\link{EmpiricalRiskMinimizationDP.CMS}} and
\code{\link{EmpiricalRiskMinimizationDP.KST}}.
}
\examples{
  coeff <- c(0.5,-1,2)
  regularizer.gr.l2(coeff)

}
\keyword{internal}
