% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{sdDP}
\alias{sdDP}
\title{Differentially Private Standard Deviation}
\usage{
sdDP(
  x,
  eps,
  lower.bounds,
  upper.bounds,
  which.sensitivity = "bounded",
  mechanism = "laplace",
  delta = NULL,
  type.DP = "pDP",
  alloc.proportions = NULL
)
}
\arguments{
\item{x}{Numeric vector, matrix, or data frame. Standard deviations taken
over columns (when applicable).}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{lower.bounds}{Numeric vector of global or public lower bounds on each
column of x. The length of lower.bounds must match the number of columns of
x (length 1 if x is a vector).}

\item{upper.bounds}{Numeric vector of global or public upper bounds on each
column of x. The length of upper.bounds must match the number of columns of
x (length 1 if x is a vector).}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively and in composition. Care must be taken not to violate
differential privacy in this case.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'laplace',
'gaussian'}. See \code{\link{LaplaceMechanism}} and
\code{\link{GaussianMechanism}} for a description of the supported
mechanisms.}

\item{delta}{Nonnegative real number defining the delta privacy parameter. If
0, reduces to eps-DP and the Laplace mechanism is used.}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism (if selected). Can be either 'pDP' for probabilistic
DP \insertCite{Liu2019a}{DPpack} or 'aDP' for approximate DP
\insertCite{DPtextbook}{DPpack}.}

\item{alloc.proportions}{Numeric vector giving the allocation proportions of
epsilon (and delta) to the statistics. For example, if this function is run
on a two-column matrix and alloc.proportions = c(.75, .25), then 75\% of the
privacy budget eps (and delta) is allocated to the statistical computation
for column 1, and the remaining 25\% is allocated to the statistical
computation for column 2. This ensures (eps, delta)-level privacy across all
computations. By default, it distributes eps and delta evenly among the
calculations. Input does not need to be normalized, meaning
alloc.proportions = c(3,1) produces the same result as the example above.}
}
\value{
A list of the sanitized standard deviations based on the bounded
and/or unbounded definitions of differential privacy.
}
\description{
This function computes the differentially private standard deviation(s) of a
given dataset at user-specified privacy levels of epsilon and delta. If the
given dataset is a matrix or data frame, differentially private standard
deviations are computed over columns and collectively satisfy differential
privacy at the specified level.
}
\examples{
sdDP(c(1,4,-2,8,-6),1,lower.bounds=-10,upper.bounds=10)
sdDP(c(1,4,-2,8,-6),1,lower.bounds=-10,upper.bounds=10,
  which.sensitivity='unbounded',mechanism='gaussian',delta=0.5,type.DP='aDP')
sdDP(matrix(c(1,4,-2,8,-6,0),ncol=2),1,lower.bounds=c(-10,-10),
  upper.bounds=c(10,10),which.sensitivity='bounded',alloc.proportions=c(1,2))

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}

\insertRef{Liu2019a}{DPpack}

\insertRef{DPtextbook}{DPpack}
}
