% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{tableDataAccess}
\alias{tableDataAccess}
\title{Differentially Private Contingency Table Data Access Function}
\usage{
tableDataAccess(x, y)
}
\arguments{
\item{x, y}{Vectors of data from which to create the contingency table.}
}
\value{
List of the true contingency table, and the bounded and unbounded
sensitivities.
}
\description{
This function performs the data access step in the computation of a
differentially private contingency table. The true values are computed using
\code{\link[base]{table}}, while the bounded and unbounded sensitivities are
computed according to the theoretical values from \link{TODO: CITE PAPER HERE}.
}
\examples{
x <- MASS::Cars93$Type;
y <- MASS::Cars93$Origin;
tableDataAccess(x,y)

}
\keyword{internal}
