% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{tableDP}
\alias{tableDP}
\title{Differentially Private Contingency Table}
\usage{
tableDP(
  x,
  y,
  eps,
  which.sensitivity = "bounded",
  mechanism = "laplace",
  delt = NULL,
  type.DP = "pDP",
  allow.negative = FALSE
)
}
\arguments{
\item{x, y}{Vectors of data from which to create the contingency table.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively. Care must be taken not to violate differential privacy in
this case.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'laplace',
'gaussian'}. See \code{\link{laplaceMechanism}} and
\code{\link{gaussianMechanism}} for a description of the supported
mechanisms.}

\item{delt}{Positive real number defining the delta privacy parameter
(necessary if using Gaussian mechanism).}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism. Can be either probabilistic DP ('pDP') or
approximate DP ('aDP'). See \code{\link{gaussianMechanism}} for a more
detailed description of this parameter. Only used if mechanism is
'gaussian'.}

\item{allow.negative}{Logical value. If FALSE (default), any negative values
in the sanitized table due to the added noise will be set to 0. If TRUE,
the negative values (if any) will be returned.}
}
\value{
A list of the sanitized contingency tables based on the bounded
and/or unbounded definitions of differential privacy.
}
\description{
This function computes a differentially private contingency table from two
vectors of data at user-specified levels of epsilon and delta.
}
\examples{
x <- MASS::Cars93$Type;
y <- MASS::Cars93$Origin;
tableDP(x,y,1,which.sensitivity='bounded',mechanism='laplace')
tableDP(x,y,1,which.sensitivity='unbounded',mechanism='gaussian',delt=0.5,
  type.DP='aDP',allow.negative=FALSE)

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}
}
