% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{histogramDataAccess}
\alias{histogramDataAccess}
\title{Differentially Private Histogram Data Access Function}
\usage{
histogramDataAccess(x, breaks)
}
\arguments{
\item{x}{Numeric vector.}

\item{breaks}{Identical to the argument with the same name from
\code{\link[graphics]{hist}}.}
}
\value{
List of the true histogram and the sensitivities calculated based on
bounded and unbounded differential privacy.
}
\description{
This function performs the data access step in the computation of a
differentially private histogram. The true values are computed using
\code{\link[graphics]{hist}}, while the sensitivities are calculated based on
bounded and unbounded differential privacy \insertCite{Kifer2011}{DPpack}
according to the theoretical values \insertCite{Liu2019b}{DPpack}.
}
\examples{
histogramDataAccess(c(1,4,3,2,3), 'Sturges')

}
\references{
\insertRef{Liu2019b}{DPpack}

\insertRef{Kifer2011}{DPpack}
}
\keyword{internal}
