% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mechanisms.R
\name{LaplaceMechanism}
\alias{LaplaceMechanism}
\title{Laplace Mechanism}
\usage{
LaplaceMechanism(
  true.values,
  eps,
  bounded.sensitivities = NULL,
  unbounded.sensitivities = NULL,
  which.sensitivity = "bounded",
  alloc.proportions = NULL
)
}
\arguments{
\item{true.values}{Real number or numeric vector corresponding to the true
value(s) of the desired statistic(s).}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{bounded.sensitivities}{Real number or numeric vector corresponding to
the global sensitivity(-ies) of the statistic based on bounded differential
privacy \insertCite{Kifer2011}{DPpack}. This is defined to be the greatest
amount by which the statistic could change in value by changing one entry
in the dataset (i.e. the total number of elements in the dataset remain the
same). This value can only be NULL if which.sensitivity is 'unbounded'.}

\item{unbounded.sensitivities}{Real number or numeric vector corresponding to
the global sensitivity(-ies) of the statistic based on unbounded
differential privacy \insertCite{Kifer2011}{DPpack}. This is defined to be
the greatest amount by which the statistic could change in value by
adding/removing one entry of the dataset (i.e. the total number of elements
in the dataset increases/decreases by one). This value can only be NULL if
which.sensitivity is 'bounded'.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively. Care must be taken not to violate differential privacy in
this case.}

\item{alloc.proportions}{Numeric vector giving the allocation proportions of
epsilon (and delta if relevant) to the statistics. For example, if
true.values is of length two and alloc.proportions = c(.75, .25), then 75\%
of the privacy budget eps (and delta) is allocated to the noise computation
for the first element of true.values, and the remaining 25\% is allocated to
the noise computation for the second element of true.values. This ensures
(eps, delta)-level privacy across all computations. By default, it
distributes eps and delta evenly among the calculations. Input does not
need to be normalized, meaning alloc.proportions = c(3,1) produces the same
result as the example above.}
}
\value{
Sanitized statistics based on the bounded and/or unbounded
definitions of differential privacy, sanitized via the Laplace mechanism.
}
\description{
This function implements the Laplace mechanism for differential privacy by
adding noise to the true value of a statistic according to specified values
of epsilon and global sensitivity. Sensitivity calculated based either on
bounded or unbounded differential privacy can be used
\insertCite{Kifer2011}{DPpack}. If true.values is a vector, the provided
epsilon is divided such that epsilon-level differential privacy is satisfied
across all statistics. If desired, the user can specify how to divide epsilon
among the statistics using alloc.proportions.
}
\examples{
LaplaceMechanism(5, 1, bounded.sensitivities=0.5,
  which.sensitivity='bounded')
LaplaceMechanism(c(5,3), 1, unbounded.sensitivities=c(1,1),
  which.sensitivity='unbounded', alloc.proportions=c(.4,.6))

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}
}
