% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{meanDataAccess}
\alias{meanDataAccess}
\title{Differentially Private Mean Data Access Function}
\usage{
meanDataAccess(x, lower.bounds, upper.bounds)
}
\arguments{
\item{x}{Numeric vector, matrix, or data frame. Means taken over columns
(when applicable).}

\item{lower.bounds}{Numeric vector of lower bounds on each column of x. The
length of lower.bounds must match the number of columns of x (length 1 if x
is a vector).}

\item{upper.bounds}{Numeric vector of upper bounds on each column of x. The
length of upper.bounds must match the number of columns of x (length 1 if x
is a vector).}
}
\value{
List of the true mean, and the bounded and unbounded sensitivities.
}
\description{
This function performs the data access step in the computation of a
differentially private mean. The true values are computed using
\code{\link[base]{mean}}, while the bounded and unbounded sensitivities are
computed according to the theoretical values from \link{TODO: CITE PAPER HERE}.
}
\examples{
meanDataAccess(c(1,4,3,2), 0, 5)

}
\keyword{internal}
