% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFunctions.R
\name{pooledVarDP}
\alias{pooledVarDP}
\title{Differentially Private Pooled Variance}
\usage{
pooledVarDP(
  ...,
  eps = 1,
  lower.bound,
  upper.bound,
  which.sensitivity = "bounded",
  mechanism = "Laplace",
  delta = 0,
  type.DP = "pDP",
  approx.n.max = FALSE
)
}
\arguments{
\item{...}{Two or more vectors from which to compute the pooled variance.}

\item{eps}{Positive real number defining the epsilon privacy budget.}

\item{lower.bound}{Real number giving the global or public lower bound of the
input data.}

\item{upper.bound}{Real number giving the global or public upper bound of the
input data.}

\item{which.sensitivity}{String indicating which type of sensitivity to use.
Can be one of {'bounded', 'unbounded', 'both'}. If 'bounded' (default),
returns result plus noise based on bounded definition for differential
privacy. If 'unbounded', returns result plus noise based on unbounded
definition. If 'both', returns result based on both methods
\insertCite{Kifer2011}{DPpack}. Note that if 'both' is chosen, each result
individually satisfies differential privacy at level eps, but may not do so
collectively and in composition. Care must be taken not to violate
differential privacy in this case.}

\item{mechanism}{String indicating which mechanism to use for differential
privacy. Currently the following mechanisms are supported: {'Laplace',
'Gaussian'}. See \code{\link{LaplaceMechanism}} and
\code{\link{GaussianMechanism}} for a description of the supported
mechanisms.}

\item{delta}{Nonnegative real number defining the delta privacy parameter. If
0 (default), reduces to eps-DP and the Laplace mechanism is used.}

\item{type.DP}{String indicating the type of differential privacy desired for
the Gaussian mechanism (if selected). Can be either 'pDP' for probabilistic
DP \insertCite{Liu2019a}{DPpack} or 'aDP' for approximate DP
\insertCite{DPtextbook}{DPpack}.}

\item{approx.n.max}{Logical indicating whether to approximate n.max, which is
defined to be the length of the largest input vector. Approximation is best
if n.max is very large.}
}
\value{
A list of the sanitized pooled variances based on the bounded and/or
unbounded definitions of differential privacy.
}
\description{
This function computes the differentially private pooled variance from two or
more vectors of data at user-specified privacy levels of epsilon and delta.
}
\examples{
pooledVarDP(c(1,4,-2,8,-6),c(1,2),c(-5,-7),eps=1,lower.bound=-10,
  upper.bound=10,which.sensitivity='bounded',mechanism='Laplace')
pooledVarDP(c(1,4,-2,8,-6),c(1,2),c(-5,-7),eps=1,
  lower.bound=-10,upper.bound=10,which.sensitivity='unbounded',
  mechanism='Gaussian',delta=0.5,type.DP='aDP',approx.n.max=TRUE)

}
\references{
\insertRef{Dwork2006a}{DPpack}

\insertRef{Kifer2011}{DPpack}

\insertRef{Liu2019a}{DPpack}

\insertRef{DPtextbook}{DPpack}
}
