% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{varDataAccess}
\alias{varDataAccess}
\title{Differentially Private Variance Data Access Function}
\usage{
varDataAccess(x, lower.bounds, upper.bounds)
}
\arguments{
\item{x}{Numeric vector, matrix, or data frame. Variances taken over columns
(when applicable).}

\item{lower.bounds}{Numeric vector of lower bounds on each column of x. The
length of lower.bounds must match the number of columns of x (length 1 if x
is a vector).}

\item{upper.bounds}{Numeric vector of upper bounds on each column of x. The
length of upper.bounds must match the number of columns of x (length 1 if x
is a vector).}
}
\value{
List of the true variance, and the bounded and unbounded
sensitivities.
}
\description{
This function performs the data access step in the computation of a
differentially private variance. The true values are computed using
\code{\link[stats]{var}}, while the bounded and unbounded sensitivities are
computed according to the theoretical values from \link{TODO: CITE PAPER HERE}.
}
\examples{
varDataAccess(c(1,4,3,2), 0, 5)

}
\keyword{internal}
